visual_language/requirements.txt
transformers
torch
Pillow
requests
groundingdino-py
supervision
segment-anything
